Move slug helper functions to helper filer and make Modal component default export
