+++ src/app/@modal/(.)products/[slug]/page.tsx
@@ -1,9 +1,7 @@
-import { Modal } from '@/components/layout/modal';
+import Modal from '@/components/layout/modal';
 import ProductCard from '@/components/product-card';
-import {
-  getProductByTitle,
-  getTitleFromSlug,
-} from '@/lib/api/products-data-server';
+import { getTitleFromSlug } from '@/lib/api/helpers';
+import { getProductByTitle } from '@/lib/api/products-data-server';
 import { notFound } from 'next/navigation';
 
 export default async function ProductModal({
