+++ src/components/merchandise-card/merchandise-card.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Button } from '@/components/ui/button';
-import { getSlugFromTitle } from '@/lib/api/products-data-server';
+import { getSlugFromTitle } from '@/lib/api/helpers';
 import { Product } from '@/lib/types/product';
 import { ImageSlider } from '../image-slider';
 import TextOutliner from '../text-outliner';
