+++ src/components/products/card-list.tsx
@@ -1,0 +1,1 @@
+export default function CardList() {}
