+++ src/lib/api/helpers.ts
@@ -1,0 +1,7 @@
+export function getSlugFromTitle(title: string): string {
+  return title.toLowerCase().replace(/\s+/g, '--');
+}
+
+export function getTitleFromSlug(slug: string): string {
+  return slug.replace(/--/g, ' ');
+}
