+++ src/lib/api/products-data-server.ts
@@ -72,11 +72,3 @@ export async function getProductsAmount(): Promise<Product[]> {
     return [];
   }
 }
-
-export function getSlugFromTitle(title: string): string {
-  return title.toLowerCase().replace(/\s+/g, '--');
-}
-
-export function getTitleFromSlug(slug: string): string {
-  return slug.replace(/--/g, ' ');
-}
